pyproject.toml
[tool.poetry]
name = "cafe-order-management-systems"
version = "0.1.0"
description = ""
authors = ["aleks <aleksejkrimsi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.5"


[tool.poetry.group.dev.dependencies]
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.2"
django-filter = "^24.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
